# openapi.yaml
openapi: 3.0.3
info:
  title: Hayaku E-commerce API
  description: Complete Next.js API backend for the Hayaku luxury e-commerce platform with Supabase integration
  version: 1.0.0
  contact:
    name: Hayaku Support
    email: support@hayaku.com
servers:
  - url: https://hayaku.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server

components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: supabase-auth-token
      description: Supabase auth token passed via cookies
  
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        short_description:
          type: string
        price:
          type: number
          format: decimal
        compare_price:
          type: number
          format: decimal
        currency:
          type: string
          default: TRY
        is_featured:
          type: boolean
        is_active:
          type: boolean
        tags:
          type: array
          items:
            type: string
        brands:
          $ref: '#/components/schemas/Brand'
        categories:
          $ref: '#/components/schemas/Category'
        product_images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
      required:
        - id
        - name
        - slug
        - price
        - currency
    
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        country:
          type: string
        is_featured:
          type: boolean
        is_active:
          type: boolean
      required:
        - id
        - name
        - slug
    
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon_name:
          type: string
        is_featured:
          type: boolean
        is_active:
          type: boolean
      required:
        - id
        - name
        - slug
    
    ProductImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        alt_text:
          type: string
        sort_order:
          type: integer
    
    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        unit_price:
          type: number
          format: decimal
        total_price:
          type: number
          format: decimal
        products:
          $ref: '#/components/schemas/Product'
      required:
        - id
        - quantity
        - unit_price
        - total_price
    
    WishlistItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        added_at:
          type: string
          format: date-time
        products:
          $ref: '#/components/schemas/Product'
      required:
        - id
        - added_at
    
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        phone:
          type: string
        vip_tier:
          type: string
          enum: [Bronze, Silver, Gold, Platinum]
        vip_points:
          type: integer
        total_orders:
          type: integer
        total_spent:
          type: number
          format: decimal
        preferred_language:
          type: string
          enum: [tr, en]
        preferred_currency:
          type: string
          default: TRY
      required:
        - id
    
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNextPage
        - hasPrevPage
    
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message description
      required:
        - error

paths:
  /products:
    get:
      summary: Get all products
      description: Retrieve products with filtering and pagination options
      tags:
        - Products
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 12
        - name: category
          in: query
          description: Filter by category slug
          schema:
            type: string
        - name: brand
          in: query
          description: Filter by brand slug
          schema:
            type: string
        - name: featured
          in: query
          description: Filter featured products
          schema:
            type: boolean
        - name: search
          in: query
          description: Search in product names/descriptions
          schema:
            type: string
        - name: minPrice
          in: query
          description: Minimum price filter
          schema:
            type: number
        - name: maxPrice
          in: query
          description: Maximum price filter
          schema:
            type: number
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            default: created_at
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create new product
      description: Create a new product (Admin only)
      tags:
        - Products
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
                price:
                  type: number
                  format: decimal
                brand_id:
                  type: string
                  format: uuid
                category_id:
                  type: string
                  format: uuid
                tags:
                  type: array
                  items:
                    type: string
              required:
                - name
                - slug
                - price
                - brand_id
                - category_id
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Product'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /products/{id}:
    get:
      summary: Get single product
      description: Retrieve a single product by ID or slug
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID or slug
          schema:
            type: string
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /brands:
    get:
      summary: Get all brands
      description: Retrieve all brands with optional filtering
      tags:
        - Brands
      parameters:
        - name: featured
          in: query
          description: Filter featured brands
          schema:
            type: boolean
        - name: active
          in: query
          description: Filter active brands
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Brands retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Brand'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /categories:
    get:
      summary: Get all categories
      description: Retrieve all categories with optional filtering
      tags:
        - Categories
      parameters:
        - name: featured
          in: query
          description: Filter featured categories
          schema:
            type: boolean
        - name: active
          in: query
          description: Filter active categories
          schema:
            type: boolean
        - name: parent
          in: query
          description: Filter by parent category ('null' for root categories)
          schema:
            type: string
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cart:
    get:
      summary: Get cart items
      description: Retrieve user's cart items
      tags:
        - Cart
      parameters:
        - name: sessionId
          in: query
          description: Required if user not authenticated
          schema:
            type: string
      responses:
        '200':
          description: Cart items retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CartItem'
        '401':
          description: Unauthorized (if no sessionId provided)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Add item to cart
      description: Add a product to the user's cart
      tags:
        - Cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  minimum: 1
                  default: 1
                sessionId:
                  type: string
                  description: Optional if user is authenticated
              required:
                - productId
      responses:
        '201':
          description: Item added to cart successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CartItem'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /wishlist:
    get:
      summary: Get wishlist items
      description: Retrieve user's wishlist items (authentication required)
      tags:
        - Wishlist
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Wishlist items retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WishlistItem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Add item to wishlist
      description: Add a product to the user's wishlist (authentication required)
      tags:
        - Wishlist
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  format: uuid
              required:
                - productId
      responses:
        '201':
          description: Item added to wishlist successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/WishlistItem'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Remove item from wishlist
      description: Remove a product from the user's wishlist (authentication required)
      tags:
        - Wishlist
      security:
        - CookieAuth: []
      parameters:
        - name: productId
          in: query
          required: true
          description: Product ID to remove
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Item removed from wishlist successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Item not found in wishlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /profile:
    get:
      summary: Get user profile
      description: Retrieve authenticated user's profile information
      tags:
        - User Profile
      security:
        - CookieAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Update user profile
      description: Update authenticated user's profile information
      tags:
        - User Profile
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                phone:
                  type: string
                birth_date:
                  type: string
                  format: date
                gender:
                  type: string
                  enum: [male, female, other]
                preferred_language:
                  type: string
                  enum: [tr, en]
                preferred_currency:
                  type: string
                  default: TRY
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserProfile'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /search:
    get:
      summary: Search across products, brands, and categories
      description: Perform a search query across different content types
      tags:
        - Search
      parameters:
        - name: q
          in: query
          required: true
          description: Search query (minimum 2 characters)
          schema:
            type: string
            minLength: 2
        - name: type
          in: query
          description: Content type to search
          schema:
            type: string
            enum: [products, brands, categories, all]
            default: all
        - name: limit
          in: query
          description: Results limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  totalResults:
                    type: integer
                  results:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
                      brands:
                        type: array
                        items:
                          $ref: '#/components/schemas/Brand'
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
        '400':
          description: Bad request (query too short)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /admin/stats:
    get:
      summary: Get dashboard statistics
      description: Retrieve admin dashboard statistics (Admin only)
      tags:
        - Admin
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      products:
                        type: object
                        properties:
                          total:
                            type: integer
                          active:
                            type: integer
                          featured:
                            type: integer
                          averagePrice:
                            type: number
                      brands:
                        type: object
                        properties:
                          total:
                            type: integer
                          active:
                            type: integer
                          featured:
                            type: integer
                      users:
                        type: object
                        properties:
                          total:
                            type: integer
                          totalRevenue:
                            type: number
                          averageOrderValue:
                            type: number
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (Admin required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /settings:
    get:
      summary: Get site settings
      description: Retrieve public site configuration settings
      tags:
        - Settings
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      site_name:
                        type: string
                      site_description:
                        type: string
                      free_shipping_threshold:
                        type: number
                      currency:
                        type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

tags:
  - name: Products
    description: Product catalog operations
  - name: Brands
    description: Brand management
  - name: Categories
    description: Category management
  - name: Cart
    description: Shopping cart operations
  - name: Wishlist
    description: Wishlist management
  - name: User Profile
    description: User account management
  - name: Search
    description: Search functionality
  - name: Admin
    description: Administrative operations
  - name: Settings
    description: Site configuration